move URL constants to better location
