+++ src/pages/Home/index.tsx
@@ -4,9 +4,9 @@ import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
 import { DAO_ROUTES } from '../../routes/constants';
 import { ReactComponent as Logo } from '../../assets/images/fractal-text-logo.svg';
-import { URL_DISCORD, URL_DOCS, URL_FAQ } from '../../components/constants';
 import { useWeb3Provider } from '../../contexts/web3Data/hooks/useWeb3Provider';
 import { ReactNode } from 'react';
+import { URL_FAQ, URL_DISCORD, URL_DOCS } from '../../constants/url';
 
 interface IconWithTextProps {
   icon: ReactNode;
